Separate swe_utils
